commit the production graph changes

+++ app/src/main/java/com/example/jetpackcomposeevoluznsewingmachine/ModalClass/CombineGraphHourDataShowing.kt
@@ -3,6 +3,6 @@ package com.example.jetpackcomposeevoluznsewingmachine.ModalClass
 data class CombineGraphHourDataShowing(
     val hour: String,
     val cycle_count: Int,
-    val total_runtime: Int,
-    val total_idleTime: Int
+    val total_runtime: Double,
+    val total_idleTime: Double
 )
