+++ app/src/main/java/com/example/jetpackcomposeevoluznsewingmachine/ModalClass/OneHourCombineGraphData.kt
@@ -2,7 +2,7 @@ package com.example.jetpackcomposeevoluznsewingmachine.ModalClass
 
  data class OneHourCombineGraphData (
     val dateTime: Int,
-    val runtime: Int,
-    val idleTime: Int,
+    val runtime: Double,
+    val idleTime: Double,
      val total_time_per_cycle:Int
  )
