+++ app/src/main/java/com/example/jetpackcomposeevoluznsewingmachine/ModalClass/SetRangeCombineGraph.kt
@@ -3,6 +3,6 @@ package com.example.jetpackcomposeevoluznsewingmachine.ModalClass
 data class SetRangeCombineGraph(
     val day:String,
    val cycle_count:Int,
-    val total_runtime:Int,
-    val total_idleTime:Int
+    val total_runtime:Double,
+    val total_idleTime:Double
 )
